% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOptions.R
\name{createOptions}
\alias{createOptions}
\title{Create a GWSDAT instance.}
\usage{
createOptions(site_name = NULL)
}
\value{
GWSDAT_Options List with essential GWSDAT options.
}
\description{
Create a GWSDAT instance.
}
