% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumeDiagnostics.R
\name{getPlumeStats}
\alias{getPlumeStats}
\title{Calculate the plume statistics including mass, area, center, and average concentration.}
\usage{
getPlumeStats(csite, substance, timestep, predicted_val, plume_thresh,
  ground_porosity)
}
\arguments{
\item{substance}{}

\item{predicted_val}{}
}
\description{
Calculate the plume statistics including mass, area, center, and average concentration.
}
