% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPSpline.R
\name{GWSDAT.compute.map.coef}
\alias{GWSDAT.compute.map.coef}
\title{Computes the coefficient vector for the MAP estimate of lambda.}
\usage{
GWSDAT.compute.map.coef(B, DtD, y, ig.a = 0.001, ig.b = 0.001, lambdas,
  prior = function(lambda) 1)
}
\description{
Computes the coefficient vector for the MAP estimate of lambda.
}
