% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPSpline.R
\name{GWSDAT.RecomputeXtinv}
\alias{GWSDAT.RecomputeXtinv}
\title{Recompute the huge Xtinv following loading a GWSDAT session.}
\usage{
GWSDAT.RecomputeXtinv(ContData, GWSDAT_Options)
}
\description{
Recompute the huge Xtinv following loading a GWSDAT session.
}
