% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch GWSDAT into server or ExcelMode.}
\usage{
launchApp(GWSDAT_Options = NULL)
}
\arguments{
\item{GWSDAT_Options}{A list of options.}
}
\description{
Launch GWSDAT into server or ExcelMode.
}
