% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pasteLimit}
\alias{pasteLimit}
\title{Convert array of strings in 'astr' to a single string separated by 'collapse', 
but only include the first 'limit' elements.}
\usage{
pasteLimit(astr, limit = NULL, collapse = ", ")
}
\description{
Convert array of strings in 'astr' to a single string separated by 'collapse', 
but only include the first 'limit' elements.
}
