% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlumeTimeSeries.R
\name{plotPlumeTimeSeries}
\alias{plotPlumeTimeSeries}
\title{Title}
\usage{
plotPlumeTimeSeries(plume_stats)
}
\arguments{
\item{plume_stats}{}
}
\description{
Title
}
\examples{
plume_stats <- getFullPlumeStats(...)
}
