% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Display concentration time-series for contaminant at specific well.}
\usage{
plotTimeSeries(csite, substance = NULL, location = NULL,
  showvline = FALSE)
}
\arguments{
\item{csite}{monitoring site object.}

\item{substance}{contaminant name}

\item{location}{well name}

\item{showvline}{show position in time (not used)}
}
\description{
Display concentration time-series for contaminant at specific well.
}
