% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Re-fit the model with the new model resolution setting, i.e. number of knots.
Update the number of knots in the text field to reflect the resolution.}
\usage{
server(input, output, session)
}
\description{
Re-fit the model with the new model resolution setting, i.e. number of knots.
Update the number of knots in the text field to reflect the resolution.
}
