% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startExcelMode.R
\name{startExcelMode}
\alias{startExcelMode}
\title{Start shiny in Excel mode.}
\usage{
startExcelMode(GWSDAT_Options = NULL)
}
\arguments{
\item{GWSDAT_Options}{A list of GWSDAT Options.}
}
\description{
Start shiny in Excel mode.
}
