% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOptions.R
\name{createOptions}
\alias{createOptions}
\title{Create start options list.}
\usage{
createOptions(site_name = NULL)
}
\arguments{
\item{site_name}{The name of the monitoring site.}
}
\value{
List containing essential model parameters and start options.
}
\description{
Create start options list.
}
\examples{
opt <- createOptions("New Site 1")
}
