% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumeDiagnostics.R
\name{getPlumeStats}
\alias{getPlumeStats}
\title{Calculate the plume statistics including mass, area, center, and average concentration for a specific time point.}
\usage{
getPlumeStats(csite, substance, timepoint, predicted_val, plume_thresh,
  ground_porosity)
}
\arguments{
\item{csite}{GWSDAT data object.}

\item{substance}{Name of the contaminant.}

\item{timepoint}{Time point (Date) for which to calculate the plume.}

\item{predicted_val}{Predicted concentration values of the contaminant.}

\item{plume_thresh}{Concentration limit defining the plume.}

\item{ground_porosity}{Porosity of the ground in percent.}
}
\value{
A data frame containing the plume statistics.
}
\description{
Calculate the plume statistics including mass, area, center, and average concentration for a specific time point.
}
