% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch GWSDAT into server or ExcelMode.}
\usage{
launchApp(GWSDAT_Options, session_file)
}
\arguments{
\item{GWSDAT_Options}{A list of start options (see createOptions).}

\item{session_file}{A .RData file with a GWSDAT analysis session.}
}
\description{
Launch GWSDAT into server or ExcelMode.
}
