% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlumeTimeSeries.R
\name{plotPlumeTimeSeries}
\alias{plotPlumeTimeSeries}
\title{Display the time-series data of a plume.}
\usage{
plotPlumeTimeSeries(plume_stats)
}
\arguments{
\item{plume_stats}{Plume statistics data frame.}
}
\description{
Display the time-series data of a plume.
}
